working on the enchantment system

+++ src/main/java/net/torocraft/jakel/loot/stat/IStatModifier.java
@@ -5,5 +5,5 @@ import net.torocraft.jakel.stats.Stats;
 @FunctionalInterface
 public interface IStatModifier {
 
-  void apply(Stats stats, int amount);
+  void apply(Stats stats, float amount);
 }
